move register_metric for calculator class and update method names
