+++ tests/sample_size/scripts/test_sample_size_run.py
@@ -13,19 +13,17 @@ class TestMain(unittest.TestCase):
         self.DEFAULT_SAMPLE_SIZE = 2000
 
     @patch("sample_size.sample_size_calculator.SampleSizeCalculator")
-    @patch("sample_size.scripts.utils.register_metric")
-    @patch("sample_size.scripts.utils.get_metric_metadata_from_input")
+    @patch("sample_size.scripts.utils.get_metric_metadata")
     @patch("sample_size.scripts.utils.get_alpha")
     def test_main_alpha_input(
         self,
         mock_get_alpha,
-        get_metric_metadata_from_input,
-        mock_register_metric,
+        get_metric_metadata,
         mock_calculator,
     ):
         test_alpha = 0.01
         mock_get_alpha.return_value = test_alpha
-        get_metric_metadata_from_input.return_value = (self.DEFAULT_METRIC_TYPE, self.DEFAULT_METRIC_METADATA)
+        get_metric_metadata.return_value = (self.DEFAULT_METRIC_TYPE, self.DEFAULT_METRIC_METADATA)
         calculator_obj = MagicMock()
         calculator_obj.get_sample_size.return_value = self.DEFAULT_SAMPLE_SIZE
         mock_calculator.return_value = calculator_obj
@@ -38,21 +36,22 @@ class TestMain(unittest.TestCase):
             )
 
         mock_get_alpha.assert_called_once()
-        get_metric_metadata_from_input.assert_called_once()
-        mock_register_metric.assert_called_once_with(
+        get_metric_metadata.assert_called_once()
+        calculator_obj.register_metric.assert_called_once_with(
             self.DEFAULT_METRIC_TYPE,
             self.DEFAULT_METRIC_METADATA,
-            calculator_obj,
         )
         calculator_obj.get_sample_size.assert_called_once()
         mock_calculator.assert_called_once_with(test_alpha)
 
     @patch("sample_size.sample_size_calculator.SampleSizeCalculator")
-    @patch("sample_size.scripts.utils.register_metric")
-    @patch("sample_size.scripts.utils.get_metric_metadata_from_input")
+    @patch("sample_size.scripts.utils.get_metric_metadata")
     @patch("sample_size.scripts.utils.get_alpha")
     def test_main_exception_print(
-        self, mock_get_alpha, get_metric_metadata_from_input, mock_register_metric, mock_calculator
+        self,
+        mock_get_alpha,
+        get_metric_metadata,
+        mock_calculator,
     ):
         error_message = "wrong alpha"
         mock_get_alpha.side_effect = Exception(error_message)
@@ -66,5 +65,4 @@ class TestMain(unittest.TestCase):
 
         mock_get_alpha.assert_called_once()
         mock_calculator.assert_not_called()
-        get_metric_metadata_from_input.assert_not_called()
-        mock_register_metric.assert_not_called()
+        get_metric_metadata.assert_not_called()
